+++ test/fixture/base-config/initConfig/plugin2.js
@@ -3,6 +3,6 @@ module.exports = function() {
 
   return {
     p2k1: {a: 1, b: 2},
-    p2k2: {c: 3, d: 4}
+    p2k2: {c: 3, d: [4]}
   };
 };