+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * fix(loot): Throw error on no such dir
 * fix(attack): Gruntfile demand() could not override modules
+* feat(event): Add `mode` argument to `grunt-horde:demand` event payload to indicate whether write was accepted
 
 # 0.2.1